fix(planning): empêcher la double déduction des plats préparés
